Implemented/Test all MVP features

+++ project1/app/src/main/java/com/revature/controllers/SpellController.java
@@ -3,6 +3,7 @@ package com.revature.controllers;
 import java.util.List;
 
 import org.apache.commons.lang3.StringUtils;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -11,13 +12,17 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.revature.exceptions.AccessDeniedException;
 import com.revature.exceptions.ItemNotFoundException;
 import com.revature.models.Spell.SpellType;
 import com.revature.models.SpellDto;
+import com.revature.models.User.UserRole;
+import com.revature.services.AuthService;
 import com.revature.services.SpellService;
 
 @RestController
@@ -25,10 +30,13 @@ import com.revature.services.SpellService;
 public class SpellController {
 	
 	private SpellService ss;
+	private AuthService as;
 	
-	public SpellController(SpellService ss) {
+	@Autowired
+	public SpellController(SpellService ss, AuthService as) {
 		super();
 		this.ss = ss;
+		this.as = as;
 	}
 	
 	@GetMapping
@@ -48,82 +56,76 @@ public class SpellController {
 		int arcCap = -1;
 		boolean queriesAllNull = true;
 		
-		if (searchStr != null && !StringUtils.isBlank(searchStr)) {
+		if (!StringUtils.isBlank(searchStr)) {
 			queriesAllNull = false;
 		}
-		if (typeStr != null) {
+		if (!StringUtils.isBlank(typeStr)) {
 			type = spellTypeFromString(typeStr);
 			if (type != SpellType.NOT_SET) queriesAllNull = false;
 		}
-		if (priceCapStr != null) {
+		if (!StringUtils.isBlank(priceCapStr)) {
 			priceCap = intFromString(priceCapStr);
 			if (priceCap >= 0) queriesAllNull = false;
 		}
-		if (inStockStr != null) {
+		if (!StringUtils.isBlank(inStockStr)) {
 			inStock = booleanFromString(inStockStr);
 			if (inStock != null) queriesAllNull = false;
 		}
-		if (intCapStr != null) {
+		if (!StringUtils.isBlank(intCapStr)) {
 			intCap = intFromString(intCapStr);
 			if (intCap >= 0) queriesAllNull = false;
 		}
-		if (faiCapStr != null) {
+		if (!StringUtils.isBlank(faiCapStr)) {
 			faiCap = intFromString(faiCapStr);
 			if (faiCap >= 0) queriesAllNull = false;
 		}
-		if (arcCapStr != null) {
+		if (!StringUtils.isBlank(arcCapStr)) {
 			arcCap = intFromString(arcCapStr);
 			if (arcCap >= 0) queriesAllNull = false;
 		}
 		
-		try {
-			List<SpellDto> spells = (queriesAllNull)
-			                            ? ss.getSpells()
-			                            : ss.getSpellsByQuery(searchStr, type, priceCap, inStock, intCap, faiCap, arcCap);
-			return new ResponseEntity<>(spells, HttpStatus.OK);
-		} catch (ItemNotFoundException e) {
-			// TODO Logging
-			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-		}
+		List<SpellDto> spells = (queriesAllNull)
+		                            ? ss.getSpells()
+		                            : ss.getSpellsByQuery(searchStr, type, priceCap, inStock, intCap, faiCap, arcCap);
+		return new ResponseEntity<>(spells, HttpStatus.OK);
 	}
 	
 	@GetMapping("/{id}")
 	public ResponseEntity<SpellDto> getSpellById(@PathVariable("id") int id) {
-		try {
 			return new ResponseEntity<>(ss.getSpellById(id), HttpStatus.OK);
-		} catch (ItemNotFoundException e) {
-			// TODO Logging
-			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-		}
 	}
 	
 	@PostMapping
-	public ResponseEntity<String> addSpell(@RequestBody SpellDto newSpell) {
+	public ResponseEntity<String> addSpell(@RequestHeader(name="Authorization", required=false) String token,
+	                                       @RequestBody SpellDto newSpell)
+	{
+		if (!as.authorizeRole(token, UserRole.STAFF, UserRole.ADMIN)) {
+			throw new AccessDeniedException("Not authorized to add spells");
+		}
 		SpellDto spell = ss.addSpell(newSpell);
-		return new ResponseEntity<>("New spell \"" + spell.name + "\" was added at index " + spell.id,
-				                    HttpStatus.CREATED);
+		return new ResponseEntity<>("New spell \"" + spell.name + "\" was added at index " + spell.id, HttpStatus.CREATED);
 	}
 	
 	@PutMapping("/{id}")
-	public ResponseEntity<SpellDto> updateSpell(@PathVariable("id") int id, @RequestBody SpellDto updates) {
-		updates.id = id;
-		try {
-			return new ResponseEntity<>(ss.updateSpell(updates), HttpStatus.CREATED);
-		} catch (ItemNotFoundException e) {
-			// TODO Logging
-			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+	public ResponseEntity<SpellDto> updateSpell(@RequestHeader(name="Authorization", required=false) String token,
+	                                            @PathVariable("id") int id, @RequestBody SpellDto updates)
+	{
+		if (!as.authorizeRole(token, UserRole.STAFF, UserRole.ADMIN)) {
+			throw new AccessDeniedException("Not authorized to modify spells");
 		}
+		updates.id = id;
+		return new ResponseEntity<>(ss.updateSpell(updates), HttpStatus.CREATED);
 	}
 	
 	@DeleteMapping("/{id}")
-	public ResponseEntity<String> deleteSpell(@PathVariable("id") int id) {
-		try {
-			SpellDto spell = ss.deleteSpell(id);
-			return new ResponseEntity<>("Deleted spell \"" + spell.name + "\"", HttpStatus.OK);
-		} catch (ItemNotFoundException e) {
-			// TODO Logging
-			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+	public ResponseEntity<String> deleteSpell(@RequestHeader(name="Authorization", required=false) String token,
+	                                          @PathVariable("id") int id)
+	{
+		if (!as.authorizeRole(token, UserRole.STAFF, UserRole.ADMIN)) {
+			throw new AccessDeniedException("Not authorized to delete spells");
 		}
+		SpellDto spell = ss.deleteSpell(id);
+		return new ResponseEntity<>("Deleted spell \"" + spell.name + "\"", HttpStatus.OK);
 	}
 	
 	private int intFromString(String intStr) {
